+++ app/(anon)/stock/[symbol]/components/StockDetail.Styled.ts
