+++ app/(anon)/stock/[symbol]/components/StockDetailTitle.tsx
@@ -7,7 +7,7 @@ import {
   StockName,
   StockPrice,
   StockDiff
-} from '@/app/(anon)/stock/[symbol]/StockDetail.Styled';
+} from '@/app/(anon)/stock/[symbol]/components/StockDetail.Styled';
 
 interface StockDetailTitleProps {
   symbol: string;
@@ -19,6 +19,7 @@ export default function StockDetailTitle({ symbol, initialPrice }: StockDetailTi
       <StockInfo>
         <StockName>{symbol}</StockName>
         <StockPrice>{initialPrice.toLocaleString()}원</StockPrice>
+        {/* TODO: 현재가, 전일대비 소켓으로 받아오기 */}
         <StockDiff>
           어제보다 <span>+315원 (1.9%)</span>
         </StockDiff>
